update code (with more comments and todos) and library.
