--- clients/vault/src/stellar_oracle/storage/mod.rs
@@ -1,2 +1,0 @@
-pub mod handler;
-pub mod traits;
